refactor(issue#65): add modifications based on folder structure discussion
