+++ src/pages/events/index.page.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import DisplayRC from '../components/EventsDisplay';
-import logo from '../public/reactdevske.svg'
+import DisplayRC from '../../components/EventsDisplay';
+import logo from '../../../public/reactdevske.svg';
 import Head from 'next/head';
 
 export const getStaticProps = async () => {
