+++ src/util/booleanFilter.js
